+++ Sources/MapboxMaps/Ornaments/DistanceFormatter.swift
@@ -11,23 +11,31 @@ internal class DistanceFormatter: MeasurementFormatter {
     ///
     /// - parameter distance: The distance, measured in meters.
     /// - returns: A localized formatted distance string including units.
-    internal func string(fromDistance distance: CLLocationDistance, useMetricSystem: Bool? = nil) -> String {
+    internal func string(fromDistance distance: CLLocationDistance, units: ScaleBarViewOptions.Units) -> String {
 
         numberFormatter.roundingIncrement = 0.25
 
         var measurement = Measurement(value: distance, unit: UnitLength.meters)
 
-        let shouldUseMetricSystem = useMetricSystem ?? locale.usesMetricSystem
-
-        if shouldUseMetricSystem {
-            unitOptions = [.providedUnit, .naturalScale]
-        } else {
+        switch units {
+        case .imperial:
             unitOptions = .providedUnit
             measurement.convert(to: .miles)
             if measurement.value <= 0.2 {
                 measurement.convert(to: .feet)
             }
+        case .nautical:
+            unitOptions = .providedUnit
+            measurement.convert(to: .nauticalMiles)
+            if measurement.value <= 0.2 {
+                measurement.convert(to: .fathoms)
+            }
+        case .metric:
+            unitOptions = [.providedUnit, .naturalScale]
+        default:
+            unitOptions = [.providedUnit, .naturalScale]
         }
+
         return string(from: measurement)
     }
 }