+++ Tests/MapboxMapsTests/Ornaments/ScaleBar/MapboxScaleBarOrnamentViewTests.swift
@@ -5,7 +5,7 @@ class MapboxScaleBarOrnamentViewTests: XCTestCase {
 
     func testImperialScaleBar() {
         let scaleBar = MockMapboxScaleBarOrnamentView()
-        scaleBar.useMetricUnits = false
+        scaleBar.units = .imperial
 
         for row in ScaleBarTestValues.imperialValues {
             scaleBar.metersPerPoint = row.metersPerPoint
@@ -17,7 +17,7 @@ class MapboxScaleBarOrnamentViewTests: XCTestCase {
 
     func testMetricScaleBar() {
         let scaleBar = MockMapboxScaleBarOrnamentView()
-        scaleBar.useMetricUnits = true
+        scaleBar.units = .metric
 
         for row in ScaleBarTestValues.metricValues {
             scaleBar.metersPerPoint = row.metersPerPoint
@@ -29,7 +29,7 @@ class MapboxScaleBarOrnamentViewTests: XCTestCase {
 
     func testImperialVisibleBars() {
             let scaleBar = MockMapboxScaleBarOrnamentView()
-            scaleBar.useMetricUnits = false
+            scaleBar.units = .imperial
 
             for row in ScaleBarTestValues.imperialValues {
                 scaleBar.metersPerPoint = row.metersPerPoint
@@ -44,7 +44,7 @@ class MapboxScaleBarOrnamentViewTests: XCTestCase {
 
         func testMetricVisibleBars() {
             let scaleBar = MockMapboxScaleBarOrnamentView()
-            scaleBar.useMetricUnits = true
+            scaleBar.units = .metric
 
             for row in ScaleBarTestValues.metricValues {
                 scaleBar.metersPerPoint = row.metersPerPoint
@@ -57,9 +57,36 @@ class MapboxScaleBarOrnamentViewTests: XCTestCase {
             }
         }
 
+    func testNauticalScaleBar() {
+        let scaleBar = MockMapboxScaleBarOrnamentView()
+        scaleBar.units = .nautical
+
+        for row in ScaleBarTestValues.nauticalValues {
+            scaleBar.metersPerPoint = row.metersPerPoint
+
+            let numberOfBars = scaleBar.preferredRow().numberOfBars
+            XCTAssertEqual(Int(numberOfBars), row.numberOfBars, "The number of scale bars should be \(row.numberOfBars) when there are \(scaleBar.metersPerPoint) feet per point.")
+        }
+    }
+
+    func testNauticalVisibleBars() {
+        let scaleBar = MockMapboxScaleBarOrnamentView()
+        scaleBar.units = .nautical
+
+        for row in ScaleBarTestValues.nauticalValues {
+            scaleBar.metersPerPoint = row.metersPerPoint
+
+            scaleBar.layoutSubviews()
+
+            let numberOfBars = row.numberOfBars
+            let visibleBars = scaleBar.dynamicContainerView.subviews
+            XCTAssertEqual(visibleBars.count, Int(numberOfBars), "\(numberOfBars) should be visible at \(row.metersPerPoint), distance: \(scaleBar.row.distance).")
+        }
+    }
+
     func testMetricScaleBarLessThanOneMeter() {
         let scaleBar = MockMapboxScaleBarOrnamentView()
-        scaleBar.useMetricUnits = true
+        scaleBar.units = .metric
 
         scaleBar.metersPerPoint = 0.004
         let preferredRow = scaleBar.preferredRow()
@@ -147,5 +174,39 @@ internal struct ScaleBarTestValues {
         (metersPerPoint: 2500.00005, numberOfBars: 2),
         (metersPerPoint: 3000.00005, numberOfBars: 3),
         (metersPerPoint: 4000.00005, numberOfBars: 2)
+    ]
+
+    // Provide test values where the metersPerPoint results in a distance slightly greater than the distance in MapboxScaleBarOrnamentView.Constants.nauticalTable
+    // These values are taken from real app zooming behavior
+    static let nauticalValues = [
+        (metersPerPoint: 0.018685636340834897, numberOfBars: 2),    // 1 fathom
+        (metersPerPoint: 0.01886784553979305, numberOfBars: 2),     // 2 fathoms
+        (metersPerPoint: 0.027432, numberOfBars: 3),                // 3 fathoms
+        (metersPerPoint: 0.04572, numberOfBars: 2),                 // 5 fathoms
+        (metersPerPoint: 0.09144, numberOfBars: 2),                 // 10 fathoms
+        (metersPerPoint: 0.18288, numberOfBars: 2),                 // 20 fathoms
+        (metersPerPoint: 0.43559919516501105, numberOfBars: 3),     // 30 fathoms
+        (metersPerPoint: 0.7195031862129788, numberOfBars: 2),      // 50 fathoms
+        (metersPerPoint: 1.4520311785905264, numberOfBars: 2),      // 100 fathoms
+        (metersPerPoint: 2.8168278728450895, numberOfBars: 2),      // 200 fathoms
+        (metersPerPoint: 11.267246571706606, numberOfBars: 2),      // 1/10 nautical mile
+        (metersPerPoint: 27.190774311218263, numberOfBars: 2),      // 1/4 nautical mile
+        (metersPerPoint: 54.8725889678557, numberOfBars: 2),        // 1/2 nautical mile
+        (metersPerPoint: 109.05781223148568, numberOfBars: 2),      // 1 nautical mile
+        (metersPerPoint: 219.45896415877314, numberOfBars: 2),      // 2 nautical miles
+        (metersPerPoint: 328.47806658042924, numberOfBars: 3),      // 3 nautical miles
+        (metersPerPoint: 547.722366413288, numberOfBars: 2),        // 5 nautical miles
+        (metersPerPoint: 1091.565564603607, numberOfBars: 2),       // 10 nautical miles
+        (metersPerPoint: 2178.148958008645, numberOfBars: 2),       // 20 nautical miles
+        (metersPerPoint: 3279.1430167603594, numberOfBars: 3),      // 30 nautical miles
+        (metersPerPoint: 5502.079305391572, numberOfBars: 2),       // 50 nautical miles
+        (metersPerPoint: 11048.75430961533, numberOfBars: 2),       // 100 nautical miles
+        (metersPerPoint: 22060.258424033687, numberOfBars: 2),      // 200 nautical miles
+        (metersPerPoint: 33191.391942996976, numberOfBars: 3),      // 300 nautical miles
+        (metersPerPoint: 52810.3283431027, numberOfBars: 2),        // 500 nautical miles
+        (metersPerPoint: 9710, numberOfBars: 2),                // 1000 nautical miles (estimated)
+        (metersPerPoint: 19144.0, numberOfBars: 2),                // 2000 nautical miles (estimated)
+        (metersPerPoint: 28976.0, numberOfBars: 3),                // 3000 nautical miles (estimated)
+        (metersPerPoint: 46546.0, numberOfBars: 2)                 // 5000 nautical miles (estimated)
     ]
 }